Split import into multiple files
